+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Current
+
+- Fix: Token equality: `TokenDetails` does a logical comparison with chain id and address,
+  instaed of object comparison. This makes TokenDetails good for ifs and hash maps. This
+  adds `TokenDetails.__eq__` and `TokenDetails.__hash__`.
+
 # 0.20
 
 - Add USDC (Centre FiatToken)